Clean select-vehicle code

+++ src/Application/Garages/Commands/CreateGarage/BriefLocationDto.cs
@@ -12,7 +12,7 @@ public class BriefLocationDto
 
     public string Country { get; set; }
 
-    public double Longitude { get; set; }
+    public float Longitude { get; set; }
 
-    public double Latitude { get; set; }
+    public float Latitude { get; set; }
 }
